backend/requirements.txt
amqp==2.2.2
apipkg==1.4
asgi-redis==1.4.3
asgiref==2.1.4
astroid==1.6.1
async-timeout==2.0.0
attrs==17.4.0
autobahn==17.10.1
Automat==0.6.0
Babel==2.5.1
bandit==1.4.0
beautifulsoup4==4.6.0
billiard==3.5.0.3
block-disposable-email==1.0.1
celery==4.1.0
certifi==2017.11.5
channels==2.0.2
chardet==3.0.4
constantly==15.1.0
coverage==4.4.2
daphne==2.0.3
defusedxml==0.5.0
Django==2.0.2
django-braces==1.11.0
django-cors-headers==2.1.0
django-disposable-email-checker==1.2.1
django-filter==1.1.0
django-oauth-toolkit==1.0.0
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.7.7
docopt==0.6.2
dodgy==0.1.9
execnet==1.5.0
factory-boy==2.9.2
Faker==0.8.10
flake8==3.5.0
flake8-polyfill==1.0.2
flower==0.9.2
gitdb2==2.0.3
GitPython==2.1.8
gunicorn==19.7.0
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
isort==4.2.15
jedi==0.11.1
kombu==4.1.0
lazy-object-proxy==1.3.1
Markdown==2.6.9
mccabe==0.6.1
mock==2.0.0
msgpack-python==0.4.8
oauthlib==2.0.6
parso==0.1.1
pbr==3.1.1
pep8-naming==0.5.0
prompt-toolkit==1.0.15
prospector==0.12.7
psycopg2==2.7.3.2
ptpython==0.41
py==1.5.2
pycodestyle==2.0.0
pydocstyle==2.1.1
pyflakes==1.6.0
Pygments==2.2.0
PyJWT==1.5.3
pylint==1.8.2
pylint-celery==0.3
pylint-common==0.2.5
pylint-django==0.9.0
pylint-flask==0.5
pylint-plugin-utils==0.2.6
pymongo==3.6.0
pytest==3.2.1
pytest-cov==2.5.1
pytest-django==3.1.2
pytest-forked==0.2
pytest-pythonpath==0.7.1
pytest-xdist==1.20.0
python-dateutil==2.6.1
python3-openid==3.1.0
pytz==2018.3
PyYAML==3.12
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
requirements-detector==0.5.2
setoptconf==0.2.0
six==1.11.0
smmap2==2.0.3
snowballstemmer==1.2.1
social-auth-app-django==2.1.0
social-auth-core==1.6.0
stevedore==1.28.0
text-unidecode==1.1
tornado==4.5.3
Twisted==17.9.0
txaio==2.8.2
urllib3==1.22
vine==1.1.4
wcwidth==0.1.7
wrapt==1.10.11
zope.interface==4.4.3
